fix: 修复代码审查发现的10项安全与逻辑问题
- 修复数据库路径硬编码，改为使用 config_get()->database_path
- 修复 API 路径越界访问，增加长度检查
- 增强静态文件路径遍历防护，拒绝 .. 及 URL 编码变体
- 修复 API 查询解析 Bug，统一使用 api_get_query_param 处理多参数
- 修复请求体读取不完整，按 Content-Length 循环读取
- 修复 Windows 下 http_server_stop 死锁（closesocket 替代 shutdown）
- 修复格式化字符串类型不匹配，int64_t 使用 PRId64
- 修复 scan_in_progress 竞态条件，增加 mutex 保护
- 修复 sscanf 返回值未检查，失败返回 400
- 修复 strncpy 边界和魔法数字问题

+++ src/modules/api_handler/api_response.c
@@ -83,13 +83,13 @@ lv_error_t api_send_json_response(int client_fd, const char *json, int status_co
     header_len = snprintf(header, sizeof(header),
                           "HTTP/1.1 %d %s\r\n"
                           "Content-Type: application/json\r\n"
-                          "Content-Length: %ld\r\n"
+                          "Content-Length: %zu\r\n"
                           "Connection: close\r\n"
                           "Access-Control-Allow-Origin: *\r\n"
                           "\r\n",
                           status_code,
                           resp.reason,
-                          (long)strlen(json));
+                          strlen(json));
 
     w1 = net_write(client_fd, header, (size_t)header_len);
     w2 = net_write(client_fd, json, strlen(json));
